Added the possibility of editing LAO name and creation time on UI
The LAO properties component in the event list now offers the possibility to edit the name of the LAO and the creation time by using the implemented edit text buttons.
This implementation uses the EdiText library from react-editext. 

NOTE - the modifications are purely on a visual level, no back-end implementation is done and thus the actual lao.name property does not change on the onSave() call of the button. For the moment, the new name or creation are simply logged on the console.

NOTE - for the moment this modification is available to all members (organiser, witnesses, attendees) which should not be the case. A future commit will allow this action only to the organiser.