Adds the ability to store/restore the organiser state
Right now the code is not decoupled and there is no use of a storage abstraction. I leave it as is for now so as to not bring too much changes at once.
It also changes the use of the channelID to always include the path, ie. with /root/... This is to avoid confusion.